util: Instantiate ACME TLS config lazy
